Move repo files to commons project and introduce lombok and map-strut
