+++ pom.xml
@@ -1,21 +1,133 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
 		 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+		<groupId>org.springframework.boot</groupId>
+		<artifactId>spring-boot-starter-parent</artifactId>
+		<version>2.7.4</version>
+		<relativePath /> <!-- lookup parent from repository -->
+	</parent>
 
 	<!-- parent pom -->
-    <groupId>org.springboardfoundation</groupId>
-    <artifactId>neednet</artifactId>
-    <packaging>pom</packaging>
+	<groupId>org.springboardfoundation</groupId>
+	<artifactId>neednet</artifactId>
+	<packaging>pom</packaging>
 	<version>1.0</version>
 
 	<!-- sub modules -->
-    <modules>
-        <module>neednet-service</module>
-        <module>users-service</module>
-        <module>needs-service</module>
-    </modules>
+	<modules>
+		<module>neednet-services-bom.xml</module>
+		<module>neednet-service</module>
+		<module>users-service</module>
+		<module>needs-service</module>
+		<module>commons</module>
+	</modules>
+
+	<properties>
+		<!-- Enable Java 11 -->
+		<java.version>11</java.version>
+		<maven.compiler.source>11</maven.compiler.source>
+		<maven.compiler.target>11</maven.compiler.target>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+
+		<!-- maven plugins -->
+		<maven.jar.plugin.version>3.2.0</maven.jar.plugin.version>
+		<maven.deploy.plugin.version>2.8.2</maven.deploy.plugin.version>
+		<maven.failsafe.plugin.version>3.0.0-M5</maven.failsafe.plugin.version>
+		<maven.surefire.plugin.version>3.0.0-M5</maven.surefire.plugin.version>
+		<maven.dependency.plugin.version>2.8</maven.dependency.plugin.version>
+		<!-- plugin properties -->
+		<lombok.version>1.18.24</lombok.version>
+		<mapstruct.version>1.4.1.Final</mapstruct.version>
+		<lombok.mapstruct.binding.version>0.2.0</lombok.mapstruct.binding.version>
+		<!-- spring dependencies -->
+		<spring.tx.version>5.3.2</spring.tx.version>
+		<spring.boot.version>2.5.13</spring.boot.version>
+		<spring-cloud.version>2021.0.3</spring-cloud.version>
+		<spring.cloud.schema.version>2.2.1.RELEASE</spring.cloud.schema.version>
+		<spring.commons.lib.version>2.12.0-SNAPSHOT</spring.commons.lib.version>
+
+		<!-- db -->
+		<mariadb.connector.version>3.0.7</mariadb.connector.version>
+		<hibernate.version>3.5.6-Final</hibernate.version>
+		<hibernate.validator.version>6.1.0.Final</hibernate.validator.version>
+
+		<!-- db -->
+		<mysql.connector.version>8.0.21</mysql.connector.version>
+		<!-- other dependencies  -->
+		<mapstruct.version>1.4.1.Final</mapstruct.version>
+		<joda.time.version>2.3</joda.time.version>
+		<springdoc.openapi.version>1.6.9</springdoc.openapi.version>
+
+		<!-- test dependencies -->
+		<junit.version>4.13.1</junit.version>
+		<groovy.version>3.0.4</groovy.version>
+		<io-rest-assured.version>4.3.1</io-rest-assured.version>
+		<wiremock.version>2.27.2</wiremock.version>
+		<awaitility.version>4.0.3</awaitility.version>
+		<badwidth.sdk.version>5.0.0</badwidth.sdk.version>
+		<badwidth.iris.sdk.version>2.0.0</badwidth.iris.sdk.version>
+		<jasypt.version>3.0.3</jasypt.version>
+
+		<jnats.version>2.11.5</jnats.version>
+		<flatbuffers.java.version>2.0.2</flatbuffers.java.version>
+		<avro.version>1.10.2</avro.version>
+		<zxing.version>3.4.1</zxing.version>
+		<log4j2.version>2.16.0</log4j2.version>
+		<json.path.version>2.6.0</json.path.version>
+		<apache.commons-text.version>1.9</apache.commons-text.version>
+	</properties>
+
+	<dependencyManagement>
+		<dependencies>
+			<dependency>
+				<groupId>org.springboardfoundation</groupId>
+				<artifactId>neednet-services-bom</artifactId>
+				<version>${project.version}</version>
+				<type>pom</type>
+				<scope>import</scope>
+			</dependency>
+		</dependencies>
+	</dependencyManagement>
+
+	<build>
+		<plugins>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<configuration>
+					<source>${java.version}</source>
+					<target>${java.version}</target>
+					<annotationProcessorPaths>
+						<path>
+							<groupId>org.mapstruct</groupId>
+							<artifactId>mapstruct-processor</artifactId>
+							<version>${mapstruct.version}</version>
+						</path>
+						<path>
+							<groupId>org.projectlombok</groupId>
+							<artifactId>lombok</artifactId>
+							<version>${lombok.version}</version>
+						</path>
+						<path>
+							<groupId>org.projectlombok</groupId>
+							<artifactId>lombok-mapstruct-binding</artifactId>
+							<version>${lombok.mapstruct.binding.version}</version>
+						</path>
+					</annotationProcessorPaths>
+				</configuration>
+			</plugin>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-failsafe-plugin</artifactId>
+			</plugin>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-deploy-plugin</artifactId>
+			</plugin>
+		</plugins>
+	</build>
 
 </project>
