+++ users-service/src/main/java/org/springboardfoundation/usersservice/controller/UserRegistrationController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@RequestMapping("/api/v1/user/register")
 @RestController
+@RequestMapping("/api/v1/user/register/")
 public class UserRegistrationController {
 
     @Autowired
