+++ users-service/src/main/resources/application.properties
@@ -1,8 +1,14 @@
-server.port=9004
-spring.datasource.url=jdbc:mariadb://localhost:3307/needs
+server.port=9001
+server.contextPath=/
+server.servlet.context-path=/
+
+spring.datasource.url=jdbc:mariadb://localhost:3306/needs
+spring.datasource.jdbcUrl=jdbc:mariadb://localhost:3306/needs
 spring.datasource.username=root
 spring.datasource.password=root
 spring.datasource.driverClassName=org.mariadb.jdbc.Driver
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=update
 spring.main.allow-bean-definition-overriding=true
+spring.jpa.database-platform=org.hibernate.dialect.MySQLDialect
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
